fixed bug in register and login

+++ src/utils/cloudinary.js
@@ -17,11 +17,18 @@ const uploadOnCloudinary = async (localFilePath) => {
 
     // console.log("file is uploaded on cloudinary", response.url);
 
-    fs.unlinkSync(localFilePath);
+    // fs.unlinkSync(localFilePath);
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
 
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath); // to remove locally saved temporary file as the upload failed so that the server wont filled with junk file(unnessary files)
+    // fs.unlinkSync(localFilePath); // to remove locally saved temporary file as the upload failed so that the server wont filled with junk file(unnessary files)
+
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
 
     return null;
   }
